# swagger/openapi .json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Online Store - OpenAPI 3.0",
    "description": "This is a backend for online store and implements:\n  - Product CRUD (include option and variants)\n  - Category CRUD\n  - Cart CRUD\n  - Order CRUD\n  \nSome useful links:\n- [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)\n- [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "https://host:port/api/v1"
    }
  ],
...
  "paths": {
...
    "/product/{productId}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Find product by ID",
        "description": "Returns a single product",
        "operationId": "getProductById",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "not found"
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
    "/admin/product": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Product List Included map of Skus",
        "description": "Can use for Admin Panel",
        "operationId": "getAdminProducts",
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "description": "Category Id that need to be considered for filter",
            "required": true,
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Default value 20",
            "required": false,
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Default value 0",
            "required": false,
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductTop"
                }
              }
            }
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "product"
        ],
        "summary": "Create a product in the store with form data",
        "description": "",
        "operationId": "createProduct",
        "requestBody": {
          "description": "Create product object",
          "content": {
            "application/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succes operation",
            "content": {
              "application/json": {
                "example": "success"
              }
            }
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      }
    },
    "/admin/product/{productId}": {
      "put": {
        "tags": [
          "product"
        ],
        "summary": "Update product by ID",
        "description": "Update product",
        "operationId": "updateProductById",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Update product",
          "content": {
            "application/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "example": "success"
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "example": "not found"
              }
            }
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "product"
        ],
        "summary": "Deletes a product",
        "description": "delete a product",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "Product id to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "example": "success"
              }
            }
          },
          "400": {
            "description": "Invalid product value"
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "example": "not found"
              }
            }
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      }
    },
    "/admin/product/{productId}/sku": {
      "post": {
        "tags": [
          "product"
        ],
        "summary": "Create a sku in the store with form data",
        "description": "",
        "operationId": "createSku",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Create sku object",
          "content": {
            "application/form-data": {
              "schema": {
                "$ref": "#/components/schemas/SkuResponse"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succes operation",
            "content": {
              "application/json": {
                "example": "success"
              }
            }
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      }
    },
    "/product/sku/{skuId}": {
      "put": {
        "tags": [
          "product"
        ],
        "summary": "Update a sku in the store with form data",
        "description": "",
        "operationId": "updateSku",
        "parameters": [
          {
            "name": "skuId",
            "in": "path",
            "description": "ID of sku",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Create sku object",
          "content": {
            "application/form-data": {
              "schema": {
                "$ref": "#/components/schemas/SkuResponse"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succes operation",
            "content": {
              "application/json": {
                "example": "success"
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "example": "not found"
              }
            }
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "product"
        ],
        "summary": "Delete a sku by productId",
        "description": "",
        "operationId": "deleteSku",
        "parameters": [
          {
            "name": "skuId",
            "in": "path",
            "description": "ID of product",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succes operation",
            "content": {
              "application/json": {
                "example": "success"
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "example": "not found"
              }
            }
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      }
    },
    "/product/skuValue/{skuValueId}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Product Options by SkuValueId",
        "description": "Options by sku value Id",
        "operationId": "getProductOption",
        "parameters": [
          {
            "name": "skuValueId",
            "in": "path",
            "description": "skuValue Id that need to be considered for filter",
            "required": true,
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SkuValueResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "example": "not found"
              }
            }
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
...
    "/admin/product/skuValue/{skuValueId}": {
      "delete": {
        "tags": [
          "product"
        ],
        "summary": "Delete Sku Option by SkuValueId",
        "description": "Options by sku value Id",
        "operationId": "deleleteProductOption",
        "parameters": [
          {
            "name": "skuValueId",
            "in": "path",
            "description": "skuValue Id that need to be considered for filter",
            "required": true,
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succes operation",
            "content": {
              "application/json": {
                "example": "success"
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "example": "not found"
              }
            }
          },
          "500": {
            "description": "internal server error"
          }
        },
        "security": [
          {
            "token_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProductResponse": {
        "required": [
          "name",
          "photoUrls"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Product Example Name"
          },
          "description": {
            "type": "string",
            "example": "Product Example Description"
          },
          "categoryId": {
            "type": "integer",
            "format": "int",
            "example": 3
          },
          "createTs": {
            "type": "string",
            "example": "2022-10-25T10:25:25.345417Z"
          },
          "countViewed": {
            "type": "integer",
            "format": "int",
            "example": 0
          },
          "skuId": {
            "type": "integer",
            "format": "int",
            "example": 1
          },
          "skuCode": {
            "type": "string",
            "example": "sku-123456"
          },
          "skuPrice": {
            "type": "integer",
            "format": "float32",
            "example": 12.3
          },
          "skuQuantity": {
            "type": "integer",
            "format": "int",
            "example": 2
          },
          "skuImage": {
            "type": "string",
            "example": "/static/sku-123456/image1.jpg, /static/sku-123456/image2.jpg"
          },
          "skuValueId": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "example": [
              1,
              2,
              3,
              4
            ]
          }
        }
      },
...
      "SkuValueResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "optionName"
          },
          "values": {
            "$ref": "#/components/schemas/OptionValueResponse"
          }
        }
      },
      "OptionValueResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "optionValueName"
          }
        }
      },
      "ProductSkuResponse": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Product Example Name"
          },
          "description": {
            "type": "string",
            "example": "Product Example Description"
          },
          "categoryId": {
            "type": "integer",
            "format": "int",
            "example": 3
          },
          "createTs": {
            "type": "string",
            "example": "2022-10-25T10:25:25.345417Z"
          },
          "variants": {
            "$ref": "#/components/schemas/SkuResponse"
          }
        }
      },
      "SkuResponse": {
        "type": "object",
        "properties": {
          "skuId": {
            "type": "integer",
            "format": "int",
            "example": 3
          },
          "skuCode": {
            "type": "string",
            "example": "sku-123456"
          },
          "skuPrice": {
            "type": "integer",
            "format": "float32",
            "example": 12.3
          },
          "skuQuantity": {
            "type": "integer",
            "format": "int",
            "example": 2
          },
          "skuImage": {
            "type": "string",
            "example": "/static/sku-123456/image1.jpg, /static/sku-123456/image2.jpg"
          }
        }
      },
      "ProductCreateReq": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "integer",
            "format": "int",
            "example": 3
          },
          "name": {
            "type": "string",
            "example": "Product Example Name"
          },
          "description": {
            "type": "string",
            "example": "Product Example Description"
          },
          "brandId": {
            "type": "integer",
            "format": "int",
            "example": 45
          },
          "regionId": {
            "type": "integer",
            "format": "int",
            "example": 33
          }
        }
      },
...
      "ProductTop": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int",
            "example": 1
          },
          "products": {
            "$ref": "#/components/schemas/ProductSkuResponse"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "token_key": {
        "type": "apiKey",
        "name": "token",
        "in": "header"
      }
    }
  }
}